modularize the server
